+++ nn/src/optimizer.rs
@@ -1,12 +1,13 @@
 use ndarray::{Array1, Array2};
 use crate::visitor::ParamVisitor;
+use crate::Float;
 
 pub struct SGD {
-    pub learning_rate: f64,
+    pub learning_rate: Float,
 }
 
 impl SGD {
-    pub fn new(learning_rate: f64) -> Self {
+    pub fn new(learning_rate: Float) -> Self {
         Self { learning_rate }
     }
 }
@@ -14,16 +15,16 @@ impl SGD {
 impl ParamVisitor for SGD {
     fn visit_array2_with_grad(
         &mut self,
-        param: &mut Array2<f64>,
-        grad: &Array2<f64>,
+        param: &mut Array2<Float>,
+        grad: &Array2<Float>,
     ) {
         *param = &*param - &(grad * self.learning_rate);
     }
 
     fn visit_array1_with_grad(
         &mut self,
-        param: &mut Array1<f64>,
-        grad: &Array1<f64>,
+        param: &mut Array1<Float>,
+        grad: &Array1<Float>,
     ) {
         *param = &*param - &(grad * self.learning_rate);
     }
